Created graphql endpoint

+++ github_action/db_cli.py
@@ -17,7 +17,6 @@ def delete_sub(sub_name):
     db_object = DatabaseObject()
     db_object.user = os.getenv("USERNAME_DB")
     db_object.password = os.getenv("PASSWORD")
-    print(db_object.password)
     db_object.host = os.getenv("HOST")
     db_object.port = os.getenv("PORT")
     db_object.database = os.getenv("DATABASE")
@@ -25,7 +24,23 @@ def delete_sub(sub_name):
     connection = db_utils.connect(db_object)
     cur = connection.cursor()
     db_utils.remove_sub(cur, sub_name, db_utils.get_table(db_object), db_object, connection)
-
+def update_sub(sub_name, dates):
+    db_object = DatabaseObject()
+    db_object.user = os.getenv("USERNAME_DB")
+    db_object.password = os.getenv("PASSWORD")
+    db_object.host = os.getenv("DBHOST")
+    db_object.port = os.getenv("PORT")
+    db_object.database = os.getenv("DATABASE")
+    db_object.table = os.getenv("TABLE")
+    print(db_object.password)
+    connection = db_utils.connect(db_object)
+    cur = connection.cursor()
+    db_utils.update_sale_date(cur, sub_name, db_utils.get_table(db_object), db_object, connection, dates)
 if __name__ == "__main__":
-    sub_name = argv[1]
-    delete_sub(sub_name)
+    sub_name = argv[2]
+    operation = argv[1]
+    if(operation == "delete"):
+        delete_sub(sub_name)
+    elif(operation == "update"):
+        dates = argv[3]
+        update_sub(sub_name, dates)
